reposync: Don't specify metadata-path
If not set, this defaults to the value of download-path, which allows the output directory to be used directly as a
clean mirror. This is needed for repositories that contain special additional metadata and/or modularity streams.

The downloaded metadata is ignored when rebuilding repositories, so this is otherwise harmless and matches what we do
with rsync.